test: all tests passed

+++ src/Lab5/ATMSystem.Application.Contracts/BankAccounts/IBankAccountService.cs
@@ -10,7 +10,7 @@ public interface IBankAccountService
     bool IsAccountExists(long id);
     BankAccount? GetAccountByIdAndPin(long id, int pinCode);
     public bool CreateAccount(long ownerId, int balance, int pin);
-    long GetBalance();
+    int GetBalance();
     WithdrawResult Withdraw(int amount);
     void Deposit(int amount);
     LoginResult Login(long id, int pinCode);
